Fix up merge
